+++ datarec/pipeline/__init__.py
@@ -1,1 +1,2 @@
 from datarec.pipeline.pipeline import Pipeline
+from datarec.pipeline.pipeline_step import PipelineStep