+++ examples/in_out/writers/sequences.py
@@ -17,15 +17,14 @@ download_example_by_url(output_folder=output_folder, filename=filename, url='htt
 import os
 print(os.path.abspath(filepath))
 
-rawdata = read_transactions_tabular(filepath, user_col='user', item_col='item', rating_col='ratings', timestamp_col='timestamp', header=0, sep='\t')
-write_sequence_tabular_inline(rawdata, os.path.join(output_folder, 'sequence_inline.tsv'), include_timestamp=True, user_col='user', sequence_col='sequence', timestamp_col='timestamp', col_sep=',', sequence_sep=' ', verbose=True)
-write_sequence_tabular_wide(rawdata, os.path.join(output_folder, 'sequence_wide.tsv'), user_col='user', col_sep=',', header=True, verbose=True)
-write_sequence_tabular_implicit(rawdata, os.path.join(output_folder, 'sequence_implicit.tsv'), col_sep=',', verbose=True)
-write_sequences_json(rawdata, os.path.join(output_folder, 'sequence.json'), include_rating=True, include_timestamp=True, item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
-write_sequences_json_items(rawdata, os.path.join(output_folder, 'sequence_items.json'), item_col='item', verbose=True)
-write_sequences_json_array(rawdata, os.path.join(output_folder, 'sequence_array.json'), include_rating=True, include_timestamp=True, user_col='user', item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
+datarec = read_transactions_tabular(filepath, user_col='user', item_col='item', rating_col='ratings', timestamp_col='timestamp', header=0, sep='\t')
+write_sequence_tabular_inline(datarec, os.path.join(output_folder, 'sequence_inline.tsv'), include_timestamp=True, user_col='user', sequence_col='sequence', timestamp_col='timestamp', col_sep=',', sequence_sep=' ', verbose=True)
+write_sequence_tabular_wide(datarec, os.path.join(output_folder, 'sequence_wide.tsv'), user_col='user', col_sep=',', header=True, verbose=True)
+write_sequence_tabular_implicit(datarec, os.path.join(output_folder, 'sequence_implicit.tsv'), col_sep=',', verbose=True)
+write_sequences_json(datarec, os.path.join(output_folder, 'sequence.json'), include_rating=True, include_timestamp=True, item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
+write_sequences_json_items(datarec, os.path.join(output_folder, 'sequence_items.json'), item_col='item', verbose=True)
+write_sequences_json_array(datarec, os.path.join(output_folder, 'sequence_array.json'), include_rating=True, include_timestamp=True, user_col='user', item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
 
-datarec = DataRec(rawdata=rawdata)
 write_sequence_tabular_inline(datarec, os.path.join(output_folder, 'sequence_inline.tsv'), include_timestamp=True, user_col='user', sequence_col='sequence', timestamp_col='timestamp', col_sep=',', sequence_sep=' ', verbose=True)
 write_sequence_tabular_wide(datarec, os.path.join(output_folder, 'sequence_wide.tsv'), user_col='user', col_sep=',', header=True, verbose=True)
 write_sequence_tabular_implicit(datarec, os.path.join(output_folder, 'sequence_implicit.tsv'), col_sep=',', verbose=True)